change folder structure: move controller/* to controller/k8s
